修正

+++ make_link.sh
@@ -7,13 +7,13 @@ if [[ ! -d $DIR ]]; then
   exit 1
 fi
 
-ln -s $DIR/.zshrc ~/.zshrc
-ln -s $DIR/.zshrc.osx ~/.zshrc.osx
-ln -s $DIR/.zsh_profile ~/.zsh_profile
-ln -s $DIR/.exrc ~/.exrc
-ln -s $DIR/.vimrc ~/.vimrc
-ln -s $DIR/.gvimrc ~/.gvimrc
-ln -s $DIR/vimfiles ~/.vim
-ln -s $DIR/.vimperatorrc ~/.vimperatorrc
-ln -s $DIR/vimperator ~/.vimperator
+[[ -e ~/.zshrc ]] || ln -s $DIR/.zshrc ~/.zshrc
+[[ -e ~/.zshrc.osx ]] || ln -s $DIR/.zshrc.osx ~/.zshrc.osx
+[[ -e ~/.zsh_profile ]] || ln -s $DIR/.zsh_profile ~/.zsh_profile
+[[ -e ~/.exrc ]] || ln -s $DIR/.exrc ~/.exrc
+[[ -e ~/.vimrc ]] || ln -s $DIR/.vimrc ~/.vimrc
+[[ -e ~/.gvimrc ]] || ln -s $DIR/.gvimrc ~/.gvimrc
+[[ -e ~/.vim ]] || ln -s $DIR/vimfiles ~/.vim
+#[[ -e ~/.vimperatorrc ]] || ln -s $DIR/.vimperatorrc ~/.vimperatorrc
+#[[ -e ~/.vimperator ]] || ln -s $DIR/vimperator ~/.vimperator
